requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
python-multipart==0.0.26
pymupdf==1.27.2.2
pypdf==6.10.2
langgraph==1.1.6
google-genai
langchain-google-genai
python-dotenv
Pillow==12.1.1
